test(docs): move quartodoc config checks into their own test module
The duplicate-contents check applies to every `docs/_quartodoc-*.yml`, not
just the playwright controller config that
`test_controller_documentation.py` is scoped to. Move it to
`test_quartodoc_configs.py` and share the config loader via
`_quartodoc_utils.py` so the two tests cannot drift.
